Work around a bug in yard(stick) / abstract_type
Yard is not able to attach a proper source_location
to a method defined by #abstract_method. However,
this only seems to be a problem if the method that
gets defined by #abstract_method is marked to be
private after the fact (as the previous code did).

In this case, yardstick/processor.rb:57:in `sort_by'
is comparing a [nil, nil] entry with other entries
of the form [String, Integer] and therefore fails
hard.